[FIX] debugging parsing, still a lot to do though :x

+++ incs/data_structures/objects.h
@@ -6,7 +6,7 @@
 /*   By: cviel <cviel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/08 18:02:40 by cviel             #+#    #+#             */
-/*   Updated: 2025/11/03 21:19:21 by cviel            ###   ########.fr       */
+/*   Updated: 2025/11/03 22:50:37 by cviel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ typedef enum e_type
 typedef struct s_sph
 {
 	t_pt3	center;
-	float	radius;
+	double	radius;
 }	t_sph;
 
 typedef struct s_plane
@@ -42,8 +42,8 @@ typedef struct s_cyl
 {
 	t_pt3	origin;
 	t_vect3	normal;
-	float	radius;
-	float	height;
+	double	radius;
+	double	height;
 }	t_cyl;
 
 typedef union u_shape
