+++ lib/libmatrix/inc/dim3.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   dim3.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mlouis <mlouis@student.42.fr>              +#+  +:+       +#+        */
+/*   By: cviel <cviel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 17:45:52 by mlouis            #+#    #+#             */
-/*   Updated: 2025/10/22 08:49:09 by mlouis           ###   ########.fr       */
+/*   Updated: 2025/11/03 22:51:45 by cviel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,14 +30,14 @@ t_mat3	mat3_trans(t_mat3 mx);
 /** VECTOR */
 typedef struct s_vect3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vect3;
 
 t_vect3	vect3_add(t_vect3 v1, t_vect3 v2);
-t_vect3	vect3_mult_nb(t_vect3 v, float nb);
-float	vect3_mult(t_vect3 v1, t_vect3 v2);
+t_vect3	vect3_mult_nb(t_vect3 v, double nb);
+double	vect3_mult(t_vect3 v1, t_vect3 v2);
 t_vect3	mat3_mult_vect3(t_mat3 mx, t_vect3 v);
 
 #endif
